1.优化了SettingsActivity和PreviewActivity之间的交互逻辑 2.修复PreviewActivity中可能存在的隐患，以及添加上错误判断，及时给予用户反馈 3.阅读方式加上选中变色

+++ app/src/main/java/com/aaron/yespdf/main/MainActivity.java
@@ -1,5 +1,6 @@
 package com.aaron.yespdf.main;
 
+import android.annotation.SuppressLint;
 import android.app.Dialog;
 import android.content.Intent;
 import android.os.Bundle;
@@ -38,11 +39,7 @@ import com.aaron.yespdf.filepicker.SelectActivity;
 import com.aaron.yespdf.settings.SettingsActivity;
 import com.blankj.utilcode.constant.PermissionConstants;
 import com.blankj.utilcode.util.ConvertUtils;
-import com.blankj.utilcode.util.KeyboardUtils;
-import com.blankj.utilcode.util.LogUtils;
 import com.blankj.utilcode.util.PermissionUtils;
-import com.blankj.utilcode.util.ThreadUtils;
-import com.github.anzewei.parallaxbacklayout.ParallaxBack;
 import com.github.mmin18.widget.RealtimeBlurView;
 import com.google.android.material.tabs.TabLayout;
 
@@ -57,16 +54,15 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.Unbinder;
 
-@ParallaxBack
-public class MainActivity extends CommonActivity implements AllAdapterComm {
+public class MainActivity extends CommonActivity implements IMainContract.V, IAllAdapterComm {
 
-    static final int SELECT_REQUEST_CODE = 101;
+    private static final int SELECT_REQUEST_CODE = 101;
 
-    //    @BindView(R2.id.app_et_search)  EditText mEtSearch;
-//    @BindView(R2.id.app_ibtn_clear) ImageButton mIbtnClear;
     @BindView(R2.id.app_tab_layout) TabLayout mTabLayout;
     @BindView(R2.id.app_vp) ViewPager mVp;
     @BindView(R2.id.app_blur_view) RealtimeBlurView mBlurView;
+
+    private IMainContract.P mP;
 
     private Unbinder mUnbinder;
     private Dialog mLoadingDialog;
@@ -76,7 +72,6 @@ public class MainActivity extends CommonActivity implements AllAdapterComm {
     private TextView mTvCollectionTitle;
     private RecyclerView mRvCollection;
     private RecyclerView.Adapter mCollectionAdapter;
-    private FragmentPagerAdapter mFragmentPagerAdapter;
 
     private List<PDF> mPDFList = new ArrayList<>();
 
@@ -90,6 +85,9 @@ public class MainActivity extends CommonActivity implements AllAdapterComm {
         return findViewById(R.id.app_toolbar);
     }
 
+    /**
+     * 只是为了在打开文档时关闭 PopupWindow
+     */
     @Subscribe(threadMode = ThreadMode.MAIN)
     public void onAllEvent(AllEvent event) {
         if (mPwCollection.isShowing()) {
@@ -102,12 +100,8 @@ public class MainActivity extends CommonActivity implements AllAdapterComm {
      */
     @Subscribe(threadMode = ThreadMode.MAIN, sticky = true)
     public void onHotfixSuccess(HotfixEvent event) {
-        hotfixRemind();
-    }
-
-    private void hotfixRemind() {
         LayoutInflater inflater = LayoutInflater.from(this);
-        View dialogView = inflater.inflate(R.layout.app_dialog_normal_alert, null);
+        View dialogView = inflater.inflate(R.layout.app_dialog_double_btn, null);
         Dialog hotfixDialog = DialogUtils.createDialog(this, dialogView);
         hotfixDialog.setCanceledOnTouchOutside(false);
         TextView tvTitle = dialogView.findViewById(R.id.app_tv_title);
@@ -128,9 +122,11 @@ public class MainActivity extends CommonActivity implements AllAdapterComm {
             @Override
             public void onViewClick(View v, long interval) {
                 final Intent intent = getPackageManager().getLaunchIntentForPackage(getPackageName());
-                intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                startActivity(intent);
-                android.os.Process.killProcess(android.os.Process.myPid());
+                if (intent != null) {
+                    intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+                    startActivity(intent);
+                    android.os.Process.killProcess(android.os.Process.myPid());
+                }
             }
         });
         hotfixDialog.show();
@@ -139,9 +135,10 @@ public class MainActivity extends CommonActivity implements AllAdapterComm {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        attachP();
         EventBus.getDefault().register(this);
         mUnbinder = ButterKnife.bind(this);
-        initView(savedInstanceState);
+        initView();
     }
 
     @Override
@@ -149,46 +146,15 @@ public class MainActivity extends CommonActivity implements AllAdapterComm {
         super.onDestroy();
         EventBus.getDefault().unregister(this);
         mUnbinder.unbind();
-        KeyboardUtils.unregisterSoftInputChangedListener(this);
+        mP.detachV();
     }
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
         if (requestCode == SELECT_REQUEST_CODE && resultCode == RESULT_OK) {
-            if (data != null) {
-                mLoadingDialog.show();
-                List<String> pathList = data.getStringArrayListExtra(SelectActivity.EXTRA_SELECTED);
-                insertPDF(pathList);
-            }
+            mP.insertPDF(data);
         }
-    }
-
-    private void insertPDF(List<String> pathList) {
-        ThreadUtils.executeByCached(new ThreadUtils.SimpleTask<Boolean>() {
-            @NonNull
-            @Override
-            public Boolean doInBackground() {
-                return DBHelper.insert(pathList);
-            }
-
-            @Override
-            public void onSuccess(Boolean success) {
-                mLoadingDialog.dismiss();
-                if (success) {
-                    UiManager.showShort(R.string.app_import_success);
-                    List<Fragment> list = getSupportFragmentManager().getFragments();
-                    for (Fragment f : list) {
-                        if (f instanceof AllFragmentComm) {
-                            ((AllFragmentComm) f).update(pathList);
-                            break;
-                        }
-                    }
-                } else {
-                    UiManager.showShort(R.string.app_import_failure);
-                }
-            }
-        });
     }
 
     @Override
@@ -218,33 +184,46 @@ public class MainActivity extends CommonActivity implements AllAdapterComm {
         mPwCollection.showAtLocation(mVp, Gravity.CENTER, 0, 0);
     }
 
-    private void initView(Bundle savedInstanceState) {
+    @Override
+    public void onShowMessage(int stringId) {
+        UiManager.showShort(stringId);
+    }
+
+    @Override
+    public void onShowLoading() {
+        mLoadingDialog.show();
+    }
+
+    @Override
+    public void onHideLoading() {
+        mLoadingDialog.dismiss();
+    }
+
+    @Override
+    public void onUpdate() {
+        List<Fragment> fragments = getSupportFragmentManager().getFragments();
+        for (Fragment f : fragments) {
+            if (f instanceof IAllFragmentComm) {
+                ((IAllFragmentComm) f).update();
+                break;
+            }
+        }
+    }
+
+    @Override
+    public void attachP() {
+        mP = new MainP(this);
+    }
+
+    private void initView() {
         mLoadingDialog = DialogUtils.createDialog(this, R.layout.app_dialog_loading);
 
         initPwMenu();
         initPwCollection();
 
-        // 监听软键盘是否打开，如果是则让 EditText 显示光标，否则不显示。
-//        KeyboardUtils.registerSoftInputChangedListener(this, height -> mEtSearch.setCursorVisible(height > 0));
-//        mEtSearch.addTextChangedListener(new TextWatcherImpl() {
-//            @Override
-//            public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-//                if (charSequence.length() != 0) {
-//                    mIbtnClear.setVisibility(View.VISIBLE);
-//                } else {
-//                    mIbtnClear.setVisibility(View.GONE);
-//                }
-//            }
-//        });
-//        mIbtnClear.setOnClickListener(v -> {
-//            mEtSearch.setText("");
-//            if (!KeyboardUtils.isSoftInputVisible(MainActivity.this)) {
-//                KeyboardUtils.showSoftInput(MainActivity.this);
-//            }
-//        });
         mTabLayout.setupWithViewPager(mVp);
-        mFragmentPagerAdapter = new MainFragmentAdapter(getSupportFragmentManager());
-        mVp.setAdapter(mFragmentPagerAdapter);
+        FragmentPagerAdapter fragmentPagerAdapter = new MainFragmentAdapter(getSupportFragmentManager());
+        mVp.setAdapter(fragmentPagerAdapter);
 
         ActionBar actionBar = getSupportActionBar();
         if (actionBar != null) {
@@ -253,6 +232,7 @@ public class MainActivity extends CommonActivity implements AllAdapterComm {
     }
 
     private void initPwMenu() {
+        @SuppressLint("InflateParams")
         View pwView = LayoutInflater.from(this).inflate(R.layout.app_pw_main, null);
         TextView tvImport   = pwView.findViewById(R.id.app_tv_import);
         TextView tvSettings = pwView.findViewById(R.id.app_tv_settings);
@@ -296,6 +276,7 @@ public class MainActivity extends CommonActivity implements AllAdapterComm {
     }
 
     private void initPwCollection() {
+        @SuppressLint("InflateParams")
         View pwView = LayoutInflater.from(this).inflate(R.layout.app_pw_collection, null);
         mTvCollectionTitle = pwView.findViewById(R.id.app_tv_title);
         mRvCollection = pwView.findViewById(R.id.app_rv_collection);
